requirements.txt
tensorflow
keras
jupyter
notebook
numpy
matplotlib
ctransformers[cuda]
Pillow
flask[async]
requests
pandas
translate
langdetect
gunicorn
youtube-search-python
websockets
flask-socketio
simple-websocket